Increase test coverage for data blobstore
